fix: HTML 페이지들의 연결이 매끄럽지 못하던 문제 수정 - /user/logout 요청 Method를 GET으로 변경한다. - /index.html 페이지에서 회원가입과 로그아웃 버튼을 동적으로 표현한다.
