+++ src/main/resources/static/index.html
@@ -16,9 +16,7 @@
             ${holder}
           </li>
           <li class="header__menu__item">
-            <a class="btn btn_ghost btn_size_s" href="/registration">
-              회원 가입
-            </a>
+            ${signupOrLogoutButton}
           </li>
         </ul>
       </header>
