+++ src/main/resources/static/user/list.html
@@ -13,12 +13,7 @@
         <a href="/"><img src="../img/signiture.svg" /></a>
         <ul class="header__menu">
             <li class="header__menu__item">
-                <a class="btn btn_contained btn_size_s" href="/login">로그인</a>
-            </li>
-            <li class="header__menu__item">
-                <a class="btn btn_ghost btn_size_s" href="/registration">
-                    회원 가입
-                </a>
+                <a class="btn btn_contained btn_size_s" href="/">홈화면으로 돌아가기</a>
             </li>
         </ul>
     </header>
