+++ src/test/java/codesquad/application/UserHandlerTest.java
@@ -48,7 +48,7 @@ class UserHandlerTest {
         HttpResponse response = userHandler.createUser(httpRequest);
 
         assertEquals(HttpStatus.FOUND, response.status());
-        assertEquals("/index.html", response.headers().get("Location"));
+        assertEquals("/", response.headers().get("Location"));
         assertEquals("유저가 생성되었습니다.", response.body());
     }
 
@@ -62,37 +62,6 @@ class UserHandlerTest {
         HttpResponse response = userHandler.createUser(httpRequest);
 
         assertEquals(HttpStatus.FOUND, response.status());
-    }
-    @Test
-    @DisplayName("홈페이지 요청을 성공적으로 처리합니다.")
-    void testGetHomepage() {
-        HttpRequest httpRequest = new HttpRequest(HttpMethod.GET, "/", new HashMap<>(), HttpProtocol.HTTP_1_1, HttpHeader.createEmpty(), new HashMap<>());
-
-        HttpResponse response = userHandler.getHomepage(httpRequest);
-
-        assertEquals(HttpStatus.OK, response.status());
-        assertTrue(response.body().contains("로그인"));
-    }
-
-    @Test
-    @DisplayName("로그인된 사용자의 홈페이지 요청을 성공적으로 처리합니다.")
-    void testGetHomepageWithLoggedInUser() {
-        // UserDatabase에 유저 추가
-        UserDatabase.addUser(testUser);
-
-        // 세션 생성 및 설정
-        Session session = sessionManager.createSession(testUser);
-        AuthenticationHolder.setContext(testUser);
-
-        Map<String, String> headers = new HashMap<>();
-        headers.put("Cookie", "SID=" + session.id());
-
-        HttpRequest httpRequest = new HttpRequest(HttpMethod.GET, "/", headers, HttpProtocol.HTTP_1_1, HttpHeader.createEmpty(), new HashMap<>());
-
-        HttpResponse response = userHandler.getHomepage(httpRequest);
-
-        assertEquals(HttpStatus.OK, response.status());
-        assertTrue(response.body().contains(testUser.getName() + "님 환영합니다."));
     }
 
     @Test
